fix(TAP-8111): Add validation to the time range of the statistical chart query. The default range has been set, and the logic for saving CPU memory indicators has been fixed

+++ manager/tm/src/main/java/com/tapdata/tm/worker/service/ApiWorkerServer.java
@@ -145,7 +145,7 @@ public class ApiWorkerServer {
         if (value instanceof Map<?, ?> infoMap) {
             Optional.ofNullable(infoMap.get("HeapMemoryUsage"))
                     .filter(Number.class::isInstance)
-                    .map(e -> ((Number) e).doubleValue())
+                    .map(e -> ((Number) e).longValue())
                     .ifPresent(info::setHeapMemoryUsage);
             Optional.ofNullable(infoMap.get("CpuUsage"))
                     .filter(Number.class::isInstance)
